Move create simple tests to new dir
And convert them to the newer test framework, and tweak a few to make use of the mutation-test action magic.
